rename device info handler file to system handler.
