+++ gantry/core/can_task.cpp
@@ -4,10 +4,10 @@
 #include "can/core/can_writer_task.hpp"
 #include "can/core/dispatch.hpp"
 #include "can/core/freertos_can_dispatch.hpp"
-#include "can/core/message_handlers/device_info.hpp"
 #include "can/core/message_handlers/motion.hpp"
 #include "can/core/message_handlers/motor.hpp"
 #include "can/core/message_handlers/move_group.hpp"
+#include "can/core/message_handlers/system.hpp"
 #include "common/core/freertos_task.hpp"
 #include "common/core/logging.hpp"
 #include "gantry/core/interfaces.hpp"
