+++ include/gantry/core/can_task.hpp
@@ -2,10 +2,10 @@
 
 #include "can/core/can_writer_task.hpp"
 #include "can/core/freertos_can_dispatch.hpp"
-#include "can/core/message_handlers/device_info.hpp"
 #include "can/core/message_handlers/motion.hpp"
 #include "can/core/message_handlers/motor.hpp"
 #include "can/core/message_handlers/move_group.hpp"
+#include "can/core/message_handlers/system.hpp"
 #include "common/core/freertos_message_queue.hpp"
 #include "gantry/core/tasks.hpp"
 #include "motor-control/core/motor.hpp"
